Separating out the three target workflows, replacing original_knowledge_source with primary_knowledge_source, and excluding evidences from the REF section
