+++ services.py
@@ -2,9 +2,14 @@ import http.client
 import json
 import logging
 import os
+import gzip
+import math
+import shutil
 from typing import Iterator
 
 from google.cloud import storage
+
+PRIMARY_KNOWLEDGE_SOURCE = "infores:text-mining-provider-targeted"
 
 
 def get_normalized_nodes(curie_list: list[str]) -> dict:  # pragma: no cover
@@ -83,7 +88,7 @@ def update_node_metadata(node: list[str], node_metadata_dict: dict, source: str)
 
     :param node: the node to add to the dictionary
     :param node_metadata_dict: the metadata dictionary
-    :param source: the original knowledge source
+    :param source: the primary knowledge source
     :returns the updated node metadata dictionary
     """
     category = node[2]
@@ -92,13 +97,13 @@ def update_node_metadata(node: list[str], node_metadata_dict: dict, source: str)
         if prefix not in node_metadata_dict[category]["id_prefixes"]:
             node_metadata_dict[category]["id_prefixes"].append(prefix)
         node_metadata_dict[category]["count"] += 1
-        node_metadata_dict[category]["count_by_source"]["original_knowledge_source"][source] += 1
+        node_metadata_dict[category]["count_by_source"]["primary_knowledge_source"][source] += 1
     else:
         node_metadata_dict[category] = {
             "id_prefixes": [prefix],
             "count": 1,
             "count_by_source": {
-                "original_knowledge_source": {
+                "primary_knowledge_source": {
                     source: 1
                 }
             }
@@ -113,18 +118,18 @@ def update_edge_metadata(edge: list, edge_metadata_dict: dict, node_dict: dict, 
     :param edge: the edge to add to the dictionary
     :param edge_metadata_dict: the metadata dictionary
     :param node_dict: the normalization dictionary
-    :param source: the original knowledge source
+    :param source: the primary knowledge source
     :returns the updated edge metadata dictionary
     """
     object_category = get_category(edge[0], normalized_nodes=node_dict)
     subject_category = get_category(edge[2], normalized_nodes=node_dict)
     triple = f"{object_category}|{edge[1]}|{subject_category}"
-    relation = edge[4]
+    relation = edge[14]
     if triple in edge_metadata_dict:
         if relation not in edge_metadata_dict[triple]["relations"]:
             edge_metadata_dict[triple]["relations"].append(relation)
         edge_metadata_dict[triple]["count"] += 1
-        edge_metadata_dict[triple]["count_by_source"]["original_knowledge_source"][source] += 1
+        edge_metadata_dict[triple]["count_by_source"]["primary_knowledge_source"][source] += 1
     else:
         edge_metadata_dict[triple] = {
             "subject": subject_category,
@@ -133,7 +138,7 @@ def update_edge_metadata(edge: list, edge_metadata_dict: dict, node_dict: dict, 
             "relations": [relation],
             "count": 1,
             "count_by_source": {
-                "original_knowledge_source": {
+                "primary_knowledge_source": {
                     source: 1
                 }
             }
@@ -148,7 +153,7 @@ def update_edge_metadata_2(edge, edge_metadata_dict: dict, node_dict: dict, sour
     :param edge: the edge to add to the dictionary
     :param edge_metadata_dict: the metadata dictionary
     :param node_dict: the normalization dictionary
-    :param source: the original knowledge source
+    :param source: the primary knowledge source
     :returns the updated edge metadata dictionary
     """
     sub = edge['subject_uniprot'] if edge['subject_uniprot'] else edge['subject_curie']
@@ -161,7 +166,7 @@ def update_edge_metadata_2(edge, edge_metadata_dict: dict, node_dict: dict, sour
         if relation not in edge_metadata_dict[triple]["relations"]:
             edge_metadata_dict[triple]["relations"].append(relation)
         edge_metadata_dict[triple]["count"] += 1
-        edge_metadata_dict[triple]["count_by_source"]["original_knowledge_source"][source] += 1
+        edge_metadata_dict[triple]["count_by_source"]["primary_knowledge_source"][source] += 1
     else:
         edge_metadata_dict[triple] = {
             "subject": subject_category,
@@ -170,7 +175,7 @@ def update_edge_metadata_2(edge, edge_metadata_dict: dict, node_dict: dict, sour
             "relations": [relation],
             "count": 1,
             "count_by_source": {
-                "original_knowledge_source": {
+                "primary_knowledge_source": {
                     source: 1
                 }
             }
@@ -230,3 +235,259 @@ def get_kgx_nodes(curies: list[str], normalized_nodes: dict[str, dict]) -> Itera
             yield [curie, name, category]
         else:
             yield []
+
+
+def get_aggregate_score(rows):
+    scores = []
+    for row in rows:
+        scores.append(get_score(row))
+    return math.fsum(scores) / float(len(scores))
+
+
+def get_score(row):
+    base_score = float(row['score'])
+    if not row['subject_idf'] or not row['object_idf']:
+        return base_score
+    else:
+        return abs(math.log10(row['subject_idf']) * math.log10(row['object_idf']) * base_score)
+
+
+def get_assertion_json(rows):
+    semmed_count = sum([row['semmed_flag'] for row in rows])
+    row1 = rows[0]
+    attributes_list = [
+        {
+            "attribute_type_id": "biolink:primary_knowledge_source",
+            "value": "infores:text-mining-provider-targeted",
+            "value_type_id": "biolink:InformationResource",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:supporting_data_source",
+            "value": "infores:pubmed",
+            "value_type_id": "biolink:InformationResource",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:has_evidence_count",
+            "value": row1['evidence_count'],
+            "value_type_id": "biolink:EvidenceCount",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:tmkp_confidence_score",
+            "value": get_aggregate_score(rows),
+            "value_type_id": "biolink:ConfidenceLevel",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:supporting_document",
+            "value": '|'.join([row['document_id'] for row in rows]),
+            "value_type_id": "biolink:Publication",
+            "attribute_source": "infores:pubmed"
+        }
+    ]
+    if semmed_count > 0:
+        attributes_list.append({
+            "attribute_type_id": "biolink:semmed_agreement_count",
+            "value": semmed_count,
+            "value_type_id": "SIO:000794",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        })
+    for row in rows:
+        attributes_list.append(get_evidence_json(row))
+    return json.dumps(attributes_list)
+
+
+def get_evidence_json(row):
+    nested_attributes = [
+        {
+            "attribute_type_id": "biolink:supporting_text",
+            "value": row['sentence'],
+            "value_type_id": "EDAM:data_3671",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:supporting_document",
+            "value": row['document_id'],
+            "value_type_id": "biolink:Publication",
+            "value_url": f"https://pubmed.ncbi.nlm.nih.gov/{str(row['document_id']).split(':')[-1]}/",
+            "attribute_source": "infores:pubmed"
+        },
+        {
+            "attribute_type_id": "biolink:supporting_text_located_in",
+            "value": row['document_zone'],
+            "value_type_id": "IAO_0000314",
+            "attribute_source": "infores:pubmed"
+        },
+        {
+            "attribute_type_id": "biolink:extraction_confidence_score",
+            "value": get_score(row),
+            "value_type_id": "EDAM:data_1772",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:subject_location_in_text",
+            "value": row['subject_span'] if row['subject_span'] else '',
+            "value_type_id": "SIO:001056",
+            "attribute_source": "infores:text-mining-provider-targeted"
+        },
+        {
+            "attribute_type_id": "biolink:object_location_in_text",
+            "value": row['object_span'] if row['object_span'] else '',
+            "value_type_id": "SIO:001056",
+            "attribute_source": "infores:text-mining-provider-targeted "
+        }
+    ]
+    if row['document_year']:
+        nested_attributes.append(
+            {
+                "attribute_type_id": "biolink:supporting_document_year",
+                "value": row['document_year'],
+                "value_type_id": "UO:0000036",
+                "attribute_source": "infores:pubmed"
+            }
+        )
+    if row['semmed_flag'] == 1:
+        nested_attributes.append(
+            {
+                "attribute_type_id": "biolink:agrees_with_data_source",
+                "value": "infores:semmeddb",
+                "value_type_id": "biolink:InformationResource",
+                "attribute_source": "infores:text-mining-provider-targeted"
+            }
+        )
+    return {
+        "attribute_type_id": "biolink:supporting_study_result",
+        "value": f"tmkp:{row['evidence_id']}",
+        "value_type_id": "biolink:TextMiningResult",
+        "value_url": f"https://tmui.text-mining-kp.org/evidence/{row['evidence_id']}",
+        "attribute_source": "infores:text-mining-provider-targeted",
+        "attributes": nested_attributes
+    }
+
+
+def get_edge(rows, predicate):
+    relevant_rows = [row for row in rows if row['predicate_curie'] == predicate]
+    if len(relevant_rows) == 0:
+        logging.debug(f'No relevant rows for predicate {predicate}')
+        return None
+    row1 = relevant_rows[0]
+    if (row1['object_curie'].startswith('PR:') and not row1['object_uniprot']) or \
+            (row1['subject_curie'].startswith('PR:') and not row1['subject_uniprot']):
+        logging.debug(f"Could not get uniprot for pr curie ({row1['object_curie']}|{row1['subject_curie']})")
+        return None
+    sub = row1['subject_uniprot'] if row1['subject_uniprot'] else row1['subject_curie']
+    obj = row1['object_uniprot'] if row1['object_uniprot'] else row1['object_curie']
+    supporting_study_results = '|'.join([f"tmkp:{row['evidence_id']}" for row in relevant_rows])
+    supporting_publications = '|'.join([row['document_id'] for row in relevant_rows])
+    qualified_predicate = ''
+    subject_aspect_qualifier = ''
+    subject_direction_qualifier = ''
+    subject_part_qualifier = ''
+    subject_form_or_variant_qualifier = ''
+    object_aspect_qualifier = ''
+    object_direction_qualifier = ''
+    object_part_qualifier = ''
+    object_form_or_variant_qualifier = ''
+    anatomical_context_qualifier = ''
+    if predicate == 'biolink:entity_positively_regulates_entity':
+        predicate = 'biolink:affects'
+        qualified_predicate = 'biolink:causes'
+        object_aspect_qualifier = 'activity_or_abundance'
+        object_direction_qualifier = 'increased'
+    elif predicate == 'biolink:entity_negatively_regulates_entity':
+        predicate = 'biolink:affects'
+        qualified_predicate = 'biolink:causes'
+        object_aspect_qualifier = 'activity_or_abundance'
+        object_direction_qualifier = 'decreased'
+    elif predicate == 'biolink:gain_of_function_contributes_to':
+        predicate = 'biolink:affects'
+        qualified_predicate = 'biolink:contributes_to'
+        subject_form_or_variant_qualifier = 'gain_of_function_variant_form'
+    elif predicate == 'biolink:loss_of_function_contributes_to':
+        predicate = 'biolink:affects'
+        qualified_predicate = 'biolink:contributes_to'
+        subject_form_or_variant_qualifier = 'loss_of_function_variant_form'
+    return [sub, predicate, obj, qualified_predicate,
+            subject_aspect_qualifier, subject_direction_qualifier,
+            subject_part_qualifier, subject_form_or_variant_qualifier,
+            object_aspect_qualifier, object_direction_qualifier,
+            object_part_qualifier, object_form_or_variant_qualifier,
+            anatomical_context_qualifier,
+            row1['assertion_id'], row1['association_curie'], get_aggregate_score(relevant_rows),
+            supporting_study_results, supporting_publications, get_assertion_json(relevant_rows)]
+
+
+def write_edges(edge_dict, output_filename):
+    logging.info("Starting edge output")
+    skipped_assertions = set([])
+    with open(output_filename, 'a') as outfile:
+        for assertion, rows in edge_dict.items():
+            predicates = set([row['predicate_curie'] for row in rows])
+            for predicate in predicates:
+                edge = get_edge(rows, predicate)
+                if not edge:
+                    skipped_assertions.add(assertion)
+                    continue
+                line = '\t'.join(str(val) for val in edge) + '\n'
+                throwaway_value = outfile.write(line)
+        outfile.flush()
+    logging.info(f'{len(skipped_assertions)} distinct assertions were skipped')
+    logging.info("Edge output complete")
+
+
+def compress(infile, outfile):
+    with open(infile, 'rb') as textfile:
+        with gzip.open(outfile, 'wb') as gzfile:
+            shutil.copyfileobj(textfile, gzfile)
+
+def decompress(infile, outfile):
+    with gzip.open(infile, 'rb') as gzfile:
+        with open(outfile, 'wb') as textfile:
+            shutil.copyfileobj(gzfile, textfile)
+
+
+def generate_metadata(edgefile, nodefile, outdir):
+    node_headers = ['id', 'name', 'category']
+    edge_headers = ['subject', 'predicate', 'object', 'qualified_predicate',
+               'subject_aspect_qualifier', 'subject_direction_qualifier',
+               'subject_part_qualifier', 'subject_form_or_variant_qualifier',
+               'object_aspect_qualifier', 'object_direction_qualifier',
+               'object_part_qualifier', 'object_form_or_variant_qualifier',
+               'anatomical_context_qualifier', 'id', 'relation', 'confidence_score',
+               'supporting_study_results', 'supporting_publications', '_attributes']
+    if not os.path.isdir(outdir):
+        os.mkdir(outdir)
+    node_metadata_dict = {}
+    node_file = os.path.join(outdir, "nodes.tsv")
+    edge_file = os.path.join(outdir, "edges.tsv")
+    metadata_file = os.path.join(outdir, "content_metadata.json")
+
+    node_lines = open(nodefile, 'r').readlines()
+    nodes = [node_line.strip().split('\t') for node_line in node_lines]
+    curies = [node[0] for node in nodes]
+    for node in nodes:
+        node_metadata_dict = update_node_metadata(node, node_metadata_dict, PRIMARY_KNOWLEDGE_SOURCE)
+    normalized_nodes = get_normalized_nodes(curies)
+    with open(node_file, 'w') as outfile:
+        tmp = outfile.write('\t'.join(node_headers) + '\n')
+        for line in node_lines:
+            tmp = outfile.write(line)
+
+    edge_metadata_dict = {}
+    with open(edgefile) as infile:
+        with open(edge_file, 'w') as outfile:
+            tmp = outfile.write('\t'.join(edge_headers) + '\n')
+            for line in infile:
+                tmp = outfile.write(line)
+                edge_metadata_dict = update_edge_metadata(line.split('\t'), edge_metadata_dict, normalized_nodes, PRIMARY_KNOWLEDGE_SOURCE)
+    metadata_dict = {
+        "nodes": node_metadata_dict,
+        "edges": list(edge_metadata_dict.values())
+    }
+    logging.info("Writing metadata file")
+    with open(metadata_file, 'w') as outfile:
+        outfile.write(json.dumps(metadata_dict))
+    logging.info("Creating tarball")
+    shutil.make_archive('targeted_assertions', 'gztar', root_dir=outdir)
